# core v5 fixes