--- a/packages/core/docs/getting-started.stories.mdx
+++ b/packages/core/docs/getting-started.stories.mdx
@@ -35,10 +35,11 @@ Preprocessor like Sass/Less or reference the CSS directly in your HTML.
 <link href="/node_modules/@cds/city/css/bundles/default.min.css" rel="stylesheet" />
 ```
 
-Add the following to your HTML to set the default Clarity body typography.
+Add the following to your HTML to set the default Clarity body typography and
+base light theme.
 
 ```html
-<body cds-text="body">
+<body cds-text="body" cds-theme="">
   ...
 </body>
 ```