--- a/packages/core/src/icon/icon.stories.mdx
+++ b/packages/core/src/icon/icon.stories.mdx
@@ -20,8 +20,51 @@ import { ClarityIcons, userIcon } from '@cds/core/icon';
 ClarityIcons.addIcons(userIcon);
 ```
 
+Icons have different accessibility requirements depending on the context they are being used in.
+There are three main use cases for icons and how they should be implemented in an
+accessible way.
+
+- Inert
+- Interactive
+- Status/Indicator
+
+### Neutral
+
+Icons by default are inert only meaning they do not provide any context
+to screen readers nor are interactive. The icon is purely cosmetic without
+additional markup.
+
 ```html
-<cds-icon shape="user"></cds-icon>
+<cds-icon shape="bars"></cds-icon>
+```
+
+### Interactive
+
+When icons are used in buttons or links they should be used in conjunction with descriptive text.
+
+```html
+<cds-button> <cds-icon shape="bars"></cds-icon> menu </cds-button>
+```
+
+If the icon in an interactive item does not have text then the parent interactive
+item (typically a button or link) should have an appropriate `aria-label` describing the interaction.
+
+```html
+<cds-icon-button aria-label="toggle menu">
+  <cds-icon shape="bars"></cds-icon>
+</cds-icon-button>
+```
+
+### Status/Indicator
+
+If an icon is used to show a status/indicator or part of a larger body of text
+content then the icon should have an `aria-label` and `role="img"` attribute.
+
+```html
+<p>
+  <cds-icon shape="exclamation-triangle" role="img" aria-label="Usage Warning"></cds-icon>
+  CPU usage is at 99% use.
+</p>
 ```
 
 If using JSX or TSX, import the icon name from `@clr/core/icon` and include it in the `shape` prop to improve type safety: