… packages/core/src/icon/icon.renderer.spec.ts
… packages/core/src/icon/utils/icon.classnames.spec.ts
… packages/core/src/icon/utils/icon.has-shape.spec.ts
… packages/core/src/icon/utils/icon.has-shape.ts
… packages/core/src/icon/utils/icon.svg-helpers.spec.ts
… packages/core/src/icon/utils/icon.svg-helpers.ts