## PR Checklist

Please check if your PR fulfills the following requirements:

- [x] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)
- [ ] If applicable, have a visual design approval

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

- [x] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, local variables)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] CI related changes
- [ ] Documentation content changes
- [ ] clarity.design website / infrastructure changes
- [ ] Other... Please describe:

## What is the new behavior?
- improve rendering performance for cds-icon. Prior we could render in-view around 10,000 instances in about 60 seconds. This updated version has this down to about 6 seconds.
- Improve a11y with icon and docs on appropriate use cases
- Added support for alert actions and cds-text links
- fixed various missing registry imports causing some components to not load 
- add missing background color from themes

Angular 11 + 10,000 icons
Before: ~57 seconds
![Screen Shot 2020-12-04 at 12 59 47 PM](https://user-images.githubusercontent.com/2021067/101533573-9f28f200-395b-11eb-8cbd-79321a5a8a08.png)

After ~7 seconds
![Screen Shot 2020-12-04 at 1 02 53 PM](https://user-images.githubusercontent.com/2021067/101533581-a18b4c00-395b-11eb-80a7-3e973284c77a.png)

![Screen Shot 2020-12-04 at 1 07 23 PM](https://user-images.githubusercontent.com/2021067/101533781-e1eaca00-395b-11eb-88fa-6d5648e8fce3.png)


## Does this PR introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this PR contains a breaking change, please describe the impact and migration path for existing applications below. -->

## Other information