@coryrylan 2020-12-14
I goofed and rebased early. I fixed the comments above but ping if there are any questions